+++ CHANGELOG.md
@@ -19,6 +19,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.1.6] - 2026-09-03
+
 ### Changed
 
 - Consolidated persistent configuration into the `pi-advice` namespace in Pi's